backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.27",
    "openai>=1.93.0",
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "fastapi>=0.115.14",
    "uvicorn>=0.34.3",
    "pydantic>=2.11.7",
    "httpx>=0.28.1",
    "langchain-tavily>=0.2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
